requirements.txt
flask
flask-wtf
flask-bootstrap
requests
pillow
numpy
matplotlib
gunicorn
WTForms
Werkzeug